Refactor KvRange
The end result is a concrete type `BoundRange` and using `Into<BoundRange>`
where functions previously took `KeyRange`. This refactoring means that errors
with invalid ranges are found earlier (in fact, they are mostly prevented statically,
where they can still occur, they do so when ranges are created, not when used)
and there is less conversion between range types in some cases. There is less
`Result` handling, and we use a named struct instead of a tuple for ranges
internally.

Signed-off-by: Nick Cameron <nrc@ncameron.org>